Fixed classpath for EclipseHelper
When passing ${project_classpath} with no argument, the current selected
project is chosen as default. When no project was selected, a error popup
would start annoying you until you clicked on some project, which depended
on scijava-common (for the EclipseHelper).

By passing the project name as argument, one assures that the classpath of
exactly that project is used.